[feature/281-fix-work-reponse-dto] refactor : WorkReadResponseDto request -> response 패키지로 이동
