+++ src/main/java/com/example/demo/service/work/WorkServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.demo.service.work;
 
 import com.example.demo.constant.ProgressStatus;
-import com.example.demo.dto.work.request.WorkReadResponseDto;
+import com.example.demo.dto.work.response.WorkReadResponseDto;
 import com.example.demo.global.exception.customexception.WorkCustomException;
 import com.example.demo.model.milestone.Milestone;
 import com.example.demo.model.project.Project;
